Merged PR 2324660: Add function to check for the status of all deployments and fail if deployment doesnt succeed
This script checks for the status of the deployment named rp-web to see if the helm installation succeeded.

For cases like Lift.Gateway we would not have the rp-web deployment.
Hence this PR modifies the script so that it first queries the names of all the deployments and then checks for the status of each deployment upto 3 attempts. If that fails then the task would fail, hence we would know that helm chart is not complete

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/2324660

Related work items: #5854272