Merged PR 2745261: Remove logging liveness probe
To know our application's health, the orchestrator need to ping our application's health API `/api/liveness-probe`. Both the k8s and the load balancer are replying this to know if our application is running.

There are lots of pings to the API. The logging of the health check is not porviding lots of value. Hence, the logging is removed.

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/2745261