Merged PR 2827609: Upgrade to support .Net Core 3.1 Worker Service
.Net Core 2 introduced the `IHostedService` generic hosting to bring all the good stuffs like dependency injection into console applications. However, there were still some features missing. So, we have some special handling logic to bring those into our hosting infra.

When we upgrade our run time to .net core 3.1, the hack logic is not working anymore.

Fortunately, .Net Core 3.0 introduced the `Worker Services`, which fill in lots of the feature gaps. We do not need to use the old hacks anymore.

This PR is removing the old hacks for .Net Core 2. This is a breaking change and require consumer to switch the console application to `Worker Services`.

This PR include some sample code for the new `Worker Services` and also how to use our infra code in it.

Documentation for local debug:

https://msazure.visualstudio.com/Liftr/_wiki/wikis/Liftr.wiki/53759/Setup-credential-for-local-debug

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/2827609