Merged PR 2855932: Add change in token manager to support certificates with subject name and issuer based authorisation
For Marketplace scenario, we need a Service Principal which is not a first party service principal. Now for this service principal, we want to associate it with a certificate that is stored in the keyvault and auto-rotated. When a cert auto rotates, its thumbprint changes, hence we cannot rely on thumbprint based authentication.

There is another way for authenticating using certs and that is based on "Subject" and "Issuer Name". So in aad we just tell that for this AAD app, trust any certificate which has this subject and issuer name.. Hence when the cert auto rotates, the thumbprint changes but issuer and subject name remain same. Hence it can be used.

Now for AAD to check if the cert is issued by the issuer and the subject, it needs the whole certificate. So we are adding a variable to the GetTokenAsync, through which when the token manager requests the token, it will send all the necessary details to AAD help with the authentication.

The manual process can be found here: https://aadwiki.windows-int.net/index.php?title=Subject_Name_and_Issuer_Authentication

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/2855932