Merged PR 2929956: Extract ObjectId from AAD token
When we do role assignment, we need to use the object id of an identity. Now, we are passing the object Id as a value in the configuration. After this change, we can extract the oid from token and just use it.

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/2929956