Merged PR 3360382: Make name and string as virtual on the ARM resource
If we dont override the property then the description is not coming in the swagger, hence it is essential to either do a new or do a override. For doing a override we need a virtual keyword in the base class.

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/3360382