Merged PR 3827437: Make each region can overwrite k8s version and bump to 1.19.3
Most AKS regions are having 1.19.0. However, west central US is not. This is providing a way to overwrite the version.

This change also bumped the aks version to 1.19.3, as it is supported by most regions now or in rolling out.

PR URL: https://msazure.visualstudio.com/DefaultCollection/Liftr/_git/Liftr.Common/pullrequest/3827437