Merged PR 4101364: Add DataAssetOptions in preparation for secret rotation
Currently, we are storing a `RPAssetOptions` object of all the connection strings in our key vault. `RPAssetOptions` contains all the primary and secondary connection strings and an active indicator. This is making the secret rotation very hard and consumer doing more work.

We are introducing a new `DataAssetOptions` class that only contains the active connection string. We do rotation at deployment time, we will renew the previous active one and swap the active one.

The rotation is not included in the change set. This CL will introduce a new contracts first. The old contracts are still kept to avoid downtime during migration.