Merged PR 4126252: Use Azure Key Vault CSI to load TLS certificate
For ECR requirments, certificate should be able to be refreshed withour doing deployment.

We are currently retrieving the TLS certificate during deployment time from key vault. Then we set the TLS cert as secret in the AKS. This means that we cannot refresh the certificate without redeployment.

In this PR, the key vault CSI (container storeage interface) is added. It will be running in the AKS cluster and retrieving the TLS certificate from key vault. Then we can rely it to auto refresh the TLS certificate.

Related work items: #7993865, #9340812