Merged PR 4134701: Change to single nuget feed
We are currently using the `Liftr` feed to publish our nuget packages. When we are consuming nuget packages, we cannot consume the `Liftr` feed, since some cache public packages are deleted (Serilog.Sinks.Console). Hence, for each repo, we are creating a new feed the add `Liftr` as the upstream feed. Each time when a new version is published to `Liftr` feed. It needs several hours to sync to the repo specific feed.

I am creating a new `LiftrOfficial` feed that does not have those pucages deleted. Hence, we can both push and read from the same feed. Then we do not need repo specific feeds.

Related work items: #9359050