Merged PR 4170930: Add Datadog Dev cert for end to end MP testing
Testing of Marketplace creation related changes is very hard as we need to first make the change in common repo and then consume in Datadog repo to test it. Adding this cert in Common repo would allow creating and deleting Marketplace resources from Common repo, hence allowing for easier testing.