Merged PR 4183999: Convert ResourceId to MarketplaceSubscription type in BillingUsageEvent
In the BillingUsageEvent the resource id represents the Marketplace subscription and should be a guid only. We should use the MarketplaceSubscription type for it so that it is clear to the user of our library that this field contains the MarketplaceSubscription information. - https://docs.microsoft.com/en-us/azure/marketplace/partner-center-portal/marketplace-metering-service-apis#metered-billing-single-usage-event

Also removing the System.Text.Serialization dependency as all our Converters are written according to Newtonsoft.json and we should use consistent library across all our serialization tasks.