Merged PR 4201212: Refactor Marketplace Rest Client into different components
This PR is a pure refactoring PR. No change is expected from Client side for it. I am doing this because I would like to add more granular exception handling for Marketplace, however MarketplaceRestClient is a very big file and adding more to it will make it bigger and difficult to understand, hence splitting it into smaller components that can easily be understood and tested in isolation as needed.