Merged PR 4217246: Exposing PurchaseException and other Marketplace Error Handling Changes
This PR is aimed to make Marketplace Error handling more robust as we are seeing lot of Purchase issues with Marketplace that needs custom handling at the RP side. Through this we want to expose such errors as first class objects from the Marketplace nuget package so that it simplifies the dev workflow for the RP owners.

Here are the high level changes:

1. **Removed creation of new exceptions at the MarketplaceARMClient and MarketplaceFulfillmentClient layer**. This would mean that these classes just log the exception but rethrow the same one. This is done so that the exceptions are exposed directly to the clients. Without this the clients will have to look at Inner Exception of MarketplaceException all the time and make decision. With this they can directly catch the exception thrown and act accordingly.

2. **Split the MarketplaceHttpException into 3 types:**
[1] RequestFailedException - Thrown when the request to Marketplace returned ErrorStatus or there was HttpRequestException.
[2] PollingFailureException - Thrown when there was an error during the Polling Operation.
[3] PurchaseFailureException - When there was an error from PurchaseFlow. For these we will also return PurchaseFailureType,  so that the RP can just look at the type and they dont have to parse the error message to convert into StatusCodes on their side.

Still need to do more work to make everything clean and also parse error messages for PurchaseException side, but raising it as a draft PR to gather early feedback.