Merged PR 4387679: Change add header to add if not exist to fix 431 issue
We recently observed issue in Whale, that when Whale is running in Production for a number of days without any new deployment, we start seeing "431 Request Header Too Long" failures when making GET request to RPaaS. When we restarted the pod the issue mitigated.

431 is due to the request header size,and HttpClient doesn't cache/reuse the request headers except the defaut ones we set (e.g. authorization header).

I saw this call in every operation in the original implementation:

_httpClient.DefaultRequestHeaders.Add(MetricTypeHeaderKey, MetricTypeHeaderValue);

This might be problem: https://docs.microsoft.com/en-us/dotnet/api/system.net.http.headers.httpheaders.add?view=net-5.0

See the remark: If the specified header is already present, values are added to the comma-separated list of values associated with the header.

It looks like we keep appending "metarp" to the same header again and again and the header length will grow unbounded over the time.