Merged PR 4431322: Allow Multiple marketplace relationship entities for a given resource id
We would like to soft delete relationship entities so that when a Webhook event is received for Unsubscribe, the request is able to reach the Webhook worker through the global Webhook web.
However, soft deleting leads to the issue that if the resource is created with the same resource id after deleting the current resource, if fails as the ResourceId key gets duplicated.

So through this PR we are allowing having single resource id to have multiple marketplace relationships.