Merged PR 4460152: Use csi to load Geneva certificates
We used to download Geneva certificates at deployment time and upload those to AKS cluster as secret. Then it will need a new deployment to rotate the certificate.
This change is using csi to load Geneva certificate. From the AKS cluster (csi), it will go to key vault to download the certificate. It will also auto-rotate new version to the AKS cluster.

![image.png](https://msazure.visualstudio.com/c98f36e6-db73-400c-b0ac-d24cf73bd450/_apis/git/repositories/584cf3c5-4198-44c1-8a8a-c63aa4da5169/pullRequests/4460152/attachments/image.png)