Merged PR 4495346: Add connection string auto-rotation
In SDL, there is a requirement for ['Secrets and certificate management'](https://msazure.visualstudio.com/DefaultCollection/One/_workitems/edit/9142306). It is requiring rotating the connection strings every 70 days. We support auto-rotation of cosmos DB and storage account connection strings every 28 days. Each time when you run the 'Infrastructure' deployment, it will check and rotate the connection strings automatically. In resource tags, it will store which key is active (primary or secondary) and a timestamp of the last rotation. When the deployment time reach the TTL (28 days), it will 1. rotate the inactive key and 2. mark the inactive one as active one. The previous active key will be marked as inactive and will be evicted from key vault secrets. Application need to pick it up later. However, the previous active connection string is still valid until the next rotation to ensure no downtime.

## SDL requirement details
Rotate your secrets [as per below](https://strikecommunity.azurewebsites.net/articles/6823/secrets-and-certificate-management.html) (automatically wherever possible), and anytime it has been exposed (e.g. you found it in source, you found it in email or in documentation, it was reported to you via a cred scanning tool etc.):
* Asymmetric keys and certificates - every 1 year
* Symmetric keys - every 1 year
* Storage account keys - every 1 years
* Connection strings - every 90 days
* Passwords - every 70 days