Merged PR 4724855: Add testability to logs
In some scenarios, we want to verify the log events are emitted. This PR add the [Serilog.Sinks.TestCorrelator](https://github.com/MitchBodmer/serilog-sinks-testcorrelator) to unit tests. After test run, it can see all the log events generated. We can use this to make sure the log events are emitted.