Merged PR 4735787: Add ACR management to VM image builder
We want to run container workload on VMSS, we need to bake the container image into the VM image. During the packer build process, we should pull down the container images and store those in VM image. To help with the ACR management process. This PR added an integrated ACR. We are providing help to create the ACR, import CDPx image to the ACR, and make the ACR credentials available in the packer build steps.