Merged PR 4905628: To avoid leak, the X509Certificate2 object needs to be disposed after usage.
To avoid leak, the X509Certificate2 object needs to be disposed after usage.
Currently we cache the cert object in CertificateStore and there is no way for it to be diposed. Neither by caller nor by callee.
To make it possible to clean up the cert object, we should not cache it. Instead we can cache the bytes but always return a new X509Certificate2 object. It's caller's responsibility to dipose after use.

Related work items: #10553234