Merged PR 5073695: Remove 'ProvisioningState' to avoid miss-usage
Previously, we were storing a 'ProvisioningState' in our own db record. However, since we are using RPaaS to store the real state. This is a duplicated value. The real value is always in RPaaS. I am seeing lots of wrong usage that leads to inconsistent bugs. This PR is removing the ProvisioningState from our own db.