Merged PR 5115071: Delete queue message on last retry
Previously, each queue message process failure will throw the message back to the queue. At dequeue time, we will check the dequeue count. When the count is more than the max dequeue count, we will delete the message.

Basically, the last dequeue is only for deleteion. However, this extra back to the queue is adding several more minutes before the actual failure. This is making our queue message processing duration inaccurate with extra back to queue time.

This PR is changing the delete to happen right after the last processing failure.