Merged PR 5208329: Retry fetch of managed identity token.
This PR adds a method to the entry-point of the SimpleDeploy app which tries 5 times to fetch a managed identity token. This is in response to deployment failures that we've seen in our Liftr deployments where the managed identity is not available.

There are two possibilities. One is that the endpoint just isn't there due to a provisioning issue, the other is that there is a timing issue and its taking some time to start-up. The idea here is to attempt to fetch it a few times (but not too long) and then fail quickly.

If we start seeing failures here in our deployments we can go back to the Ev2 team and get their assistance.

@<Wuyi Weng> this is my follow-up to our conversation on Teams.