Merged PR 5239330: Configure retry policy for MSI.
The Azure Grafana service deployments appear to be failing because by the time the application has started up and is requesting a managed identity, the underlying AAD pod that supplies the IDMS endpoint is still spinning up. To overcome this issue, this PR adds retry policy settings to the ```ManagedIdentityCredential``` so that it retries for up to a minute before giving up.

This mimics some of the retry behavior we've used on deployment for Ev2 shells which share some of the same technical underpinnings.

Related work items: #11930000