Merged PR 5238205: Adding additional field createGlobalDBWithZoneRedundancy to make the global cosmos db "non zone redundant"
### What does this PR do?
Adding additional field createGlobalDBWithZoneRedundancy in the global section of hosting-options.json make the global cosmos db "non zone redundant"

### Background
Basically, adding an extra option in the hosting options global section to .. Currently, if we create the new cosmos db with zone redudancy turned ON, then it fails while creating the collection with the error saying there is no capacity. So, we got the recommendation from the cosmos db capacity team to create the db without zonal redundancy for now, and later when there is capacity, they will support the migration.

### What all main changes does this PR have?
1. Introduces the new parameter createGlobalDBWithZoneRedundancy in the GlobalOptions with default value true.
2. During Global Deployment, passes this extra parameter by reading it in hosting-options.json

### What all testing are done?
1. Added an integration test to create db without zonal redundancy and then verifying it. (Skipped this test since the cosmos db tests are flaky)

Related work items: #12416606