Merged PR 5270099: Grafana IcM connector
This PR is adding the Grafana alert to IcM connector. At the time that Grafana generate an alert, it can use webhook to send it to other alert sink. This PR is adding a new REST API to receive this webhook.

To set the severity, users need to set a tag in the alert message, e.g. [sev3].

![image.png](https://msazure.visualstudio.com/c98f36e6-db73-400c-b0ac-d24cf73bd450/_apis/git/repositories/584cf3c5-4198-44c1-8a8a-c63aa4da5169/pullRequests/5270099/attachments/image.png)