Merged PR 5686731: Serilize traffic manager related functional tests
There is a new traffic manager quota. It is like 3 TM per subscription. When we are running the TM related test concurrently, we will hit the 3 TM quota very easily. This PR is serilizing all the TM related tests to meet the quota.

Marketplace mock API is broken. Akshita is having another work item to track fixing those mp tests.

![image.png](https://msazure.visualstudio.com/c98f36e6-db73-400c-b0ac-d24cf73bd450/_apis/git/repositories/584cf3c5-4198-44c1-8a8a-c63aa4da5169/pullRequests/5686731/attachments/image.png)