Merged PR 5695393: Change nuget packge to avoid leagl check
Change nuget packge to avoid leagl check

Root dependencies for Oracle.ManagedDataAccess.Core
DistributedLock
2.3.0

To validate the use of open source, Component Governance needs additional license information. Our tooling depends on the publicly accessible open source database ClearlyDefined for this information, and ClearlyDefined does not have information for this component. This is a rare occurrence, and requires input from you.

Please provide:

The license for this component;
A url to the license for this component;
Optionally, describe why you think this is the license for this component.
When you click "submit" Component Governance will automatically create a pull request in ClearlyDefined for the ClearlyDefined open source community to review according to these guidelines - ClearlyDefined Review Guidelines.

With your submission please keep in mind ClearlyDefined community members are not Microsoft employees, and donate their time when reviewing submissions.