Merged PR 5749231: Remove logger being passed into Metric Startup Extensions
Currently in Nginx Build Pipeline we are getting errors:

```
Unhandled exception. System.InvalidOperationException: Unable to resolve service for type 'Serilog.ILogger' while attempting to activate 'Microsoft.Liftr.Nginx.RP.Web.Startup'
```
This is happening because we are having logger in the constructor of the Startup class, which I think doesnt get initialized during Build time and causes the error.

Due to this, we are not able to generate the swagger file when we are building RP.Web. We would like to add a test in Nginx which compares the newly generated swagger file from changes in RP.Web to one that we have checked in. This will help us to add reviewers for the swagger file and if changes are added there, the reviewer can review . Here is the PR in Nginx -

[Pull Request 5662186: [WIP] Add test which will fail when there is a swagger change](https://msazure.visualstudio.com/Liftr/_git/Liftr.Nginx/pullrequest/5662186)

The logger in the constructor is needed to pass to this method in the StartupExtensions. This method doesn't need logger to be passed, it can read it from the Service Provider like all our extension methods are doing. Its a breaking change but shouldnt have much impact. I will notify on the common channel once this change is merged.