Merged PR 5790535: [VNet Injection] Add models and Datasource to be used by RP for VNet Injection
VNet Injection has been under development in the Liftr.Nginx repo and is currently working well for Nginx Partner. Now we would like to extend it to other partners like Qumulo by adding the common components in the Common repo and the actual VNet Injection service in the Gateway repo.

The projects added in this PR are the ones that Liftr RP will use to add the VNet Injection Entity to the database which will be consumed by the VNet Injection Service to perform validations on the incoming requests.

Related work items: #13186197