Merged PR 5991355: [Bug Fix] correctly check existence of ACR ImageScanEventSubscription
## Context
In infrastructure rollout, we check if an ACR is associated to ImageScanEventSubscription. If not, we will create the EventGrid Subscription for the ACR.

## Bug
When checking the existence of event subscription, I used the wrong `scope` in
` eventGridManagementClient.EventSubscriptions.GetWithHttpMessagesAsync(scope, eventSubscriptionName)`

## Impact
The impact is, we always create a ImageScanEventSubscription for each ACR during infra deployment, even when the event subscription already exists.
![image.png](https://msazure.visualstudio.com/c98f36e6-db73-400c-b0ac-d24cf73bd450/_apis/git/repositories/584cf3c5-4198-44c1-8a8a-c63aa4da5169/pullRequests/5991355/attachments/image.png)

## Fix
The correct `scope` should be the resourceId of the target ACR, i.e.
```
var scope = $"/subscriptions/{DefaultSubscriptionId}/resourceGroups/{acr.ResourceGroupName}/providers/Microsoft.ContainerRegistry/registries/{acr.Name}";
```

## Verification
Verified in `Liftr.Fluent.Tests.AcrTests.cs`